# Split merge and compare